[REFACTOR/#27] 전체 코드 수정, 대표 뷰 파일화, 세부 뷰 귀속
