+++ app/src/main/java/com/sopt/bubble/feature/precise_store/PreciseStoreScreen.kt
@@ -1,9 +1,5 @@
 package com.sopt.bubble.feature.precise_store
 
-import androidx.compose.animation.animateContentSize
-import androidx.compose.animation.core.Spring
-import androidx.compose.animation.core.spring
-import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Spacer
@@ -12,22 +8,16 @@ import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.wrapContentSize
 import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.verticalScroll
 import androidx.compose.material3.Scaffold
-import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
@@ -37,26 +27,15 @@ import androidx.navigation.compose.rememberNavController
 import coil.compose.AsyncImage
 import coil.request.ImageRequest
 import com.sopt.bubble.R
-import com.sopt.bubble.data.dto.response.ResponsePreciseArtistDto
-import com.sopt.bubble.feature.precise_store.PreciseStoreViewModel.Companion.CHECK_BUTTON_NUM
-import com.sopt.bubble.feature.precise_store.PreciseStoreViewModel.Companion.MORE_UNFOLD_ITEM_LIMIT
-import com.sopt.bubble.feature.precise_store.PreciseStoreViewModel.Companion.PRECISE_STORE_BANNER_IMAGE_RATIO
 import com.sopt.bubble.feature.precise_store.PreciseStoreViewModel.Companion.PRECISE_STORE_TOP_IMAGE_RATIO
+import com.sopt.bubble.feature.precise_store.component.PreciseMoreSubscribeView
+import com.sopt.bubble.feature.precise_store.component.PreciseStoreArtistDescriptionView
 import com.sopt.bubble.feature.precise_store.component.PreciseStoreBottomBar
-import com.sopt.bubble.feature.precise_store.component.PreciseStoreCheckBox
-import com.sopt.bubble.feature.precise_store.component.PreciseStoreMoreButton
-import com.sopt.bubble.feature.precise_store.component.PreciseStoreTicket
+import com.sopt.bubble.feature.precise_store.component.PreciseStoreBubbleDescriptionView
+import com.sopt.bubble.feature.precise_store.component.PreciseStoreCheckBoxView
 import com.sopt.bubble.feature.precise_store.component.PreciseStoreTopBar
-import com.sopt.bubble.feature.precise_store.model.checkBoxList
-import com.sopt.bubble.ui.theme.Body02
-import com.sopt.bubble.ui.theme.Body03
-import com.sopt.bubble.ui.theme.Gray300
-import com.sopt.bubble.ui.theme.Gray400
-import com.sopt.bubble.ui.theme.Gray500
 import com.sopt.bubble.ui.theme.Gray700
 import com.sopt.bubble.ui.theme.Gray800
-import com.sopt.bubble.ui.theme.Headline04
-import com.sopt.bubble.ui.theme.White
 
 @Composable
 fun PreciseStoreScreen(
@@ -103,19 +82,6 @@ fun PreciseStoreScreen(
 }
 
 @Composable
-fun PreciseStoreFailureScreen(
-    modifier: Modifier = Modifier
-) {
-    Column(
-        modifier = modifier
-            .fillMaxSize()
-            .background(Gray700)
-    ) {
-
-    }
-}
-
-@Composable
 fun PreciseStoreSuccessScreen(
     modifier: Modifier = Modifier,
     uiState: PreciseStoreState.SuccessState,
@@ -150,7 +116,7 @@ fun PreciseStoreSuccessScreen(
                     .aspectRatio(PRECISE_STORE_TOP_IMAGE_RATIO)
             )
 
-            PreciseStoreArtistDescription(
+            PreciseStoreArtistDescriptionView(
                 artistName = uiState.name,
                 bubbleDescription = uiState.bubbleDescription,
                 serviceMember = uiState.isServiceMember,
@@ -160,7 +126,7 @@ fun PreciseStoreSuccessScreen(
                     .padding(top = 16.dp)
             )
 
-            PreciseMoreView(subscribeList = uiState.subscribes)
+            PreciseMoreSubscribeView(subscribeList = uiState.subscribes)
 
             Spacer(
                 modifier = Modifier
@@ -169,7 +135,7 @@ fun PreciseStoreSuccessScreen(
                     .background(color = Gray800)
             )
 
-            PreciseStoreBubbleDescription(
+            PreciseStoreBubbleDescriptionView(
                 description = uiState.description,
                 modifier = Modifier
                     .padding(horizontal = 20.dp)
@@ -177,7 +143,7 @@ fun PreciseStoreSuccessScreen(
             )
 
 
-            PreciseStoreCheckBoxes(
+            PreciseStoreCheckBoxView(
                 onClickCheckBox = { onClickCheckBox(it) },
                 isCheckedList = uiState.isCheckedList,
                 modifier = Modifier
@@ -192,169 +158,18 @@ fun PreciseStoreSuccessScreen(
 }
 
 @Composable
-private fun PreciseStoreArtistDescription(
-    artistName: String,
-    bubbleDescription: String,
-    serviceMember: String,
-    nonServiceMember: String,
+fun PreciseStoreFailureScreen(
     modifier: Modifier = Modifier
 ) {
     Column(
         modifier = modifier
+            .fillMaxSize()
+            .background(Gray700)
     ) {
-        /*아티스트 이름 텍스트*/
-        Text(
-            text = artistName.ifEmpty { "" },
-            color = White,
-            style = Headline04,
-            modifier = Modifier.padding(top = 16.dp)
-        )
 
-        Spacer(
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(top = 16.dp)
-                .height(1.dp)
-                .background(color = Gray800)
-        )
-
-        /*아티스트 버블 소개 텍스트*/
-        Text(
-            text = bubbleDescription,
-            color = Gray300,
-            style = Body03,
-            modifier = Modifier.padding(top = 20.dp)
-        )
-
-        /*아티스트 라인업 텍스트*/
-        Text(
-            text = stringResource(id = R.string.precise_store_artist_lineup),
-            color = White,
-            style = Body02,
-            modifier = Modifier.padding(top = 18.dp),
-        )
-
-        Text(
-            text = serviceMember,
-            color = White,
-            style = Body03,
-            modifier = Modifier.padding(top = 6.dp)
-        )
-
-        /*아티스트 커밍순 텍스트*/
-        if (nonServiceMember.isNotBlank()) {
-            Text(
-                text = stringResource(id = R.string.precise_store_artist_coming_soon),
-                color = Gray500,
-                style = Body02,
-                modifier = Modifier.padding(top = 18.dp)
-            )
-            Text(
-                text = nonServiceMember,
-                color = Gray500,
-                style = Body03,
-                modifier = Modifier.padding(top = 6.dp)
-            )
-        }
     }
 }
 
-@Composable
-private fun PreciseMoreView(
-    subscribeList: List<ResponsePreciseArtistDto.Result.Artist.Subscribe>
-) {
-    var isMorePressed by remember { mutableStateOf(false) }
-    val moreIndex =
-        if (subscribeList.size < MORE_UNFOLD_ITEM_LIMIT || isMorePressed) subscribeList.size
-        else MORE_UNFOLD_ITEM_LIMIT
-
-    Column(
-        modifier = Modifier
-            .wrapContentSize()
-            .animateContentSize(
-                animationSpec = spring(
-                    dampingRatio = Spring.DampingRatioNoBouncy,
-                    stiffness = Spring.StiffnessMedium
-                )
-            )
-    ) {
-        for (subscribe in subscribeList.subList(0, moreIndex)) {
-            PreciseStoreTicket(
-                title = subscribe.name,
-                price = subscribe.price,
-                originalPrice = subscribe.previousPrice,
-                modifier = if (subscribeList.indexOf(subscribe) != 0) Modifier.padding(top = 14.dp)
-                else Modifier.padding(top = 26.dp)
-            )
-        }
-    }
-
-    if (!isMorePressed && subscribeList.size > MORE_UNFOLD_ITEM_LIMIT) {
-        PreciseStoreMoreButton(onClick = { isMorePressed = !isMorePressed })
-    } else {
-        Spacer(modifier = Modifier.height(26.dp))
-    }
-}
-
-@Composable
-private fun PreciseStoreBubbleDescription(
-    description: String,
-    modifier: Modifier = Modifier
-) {
-    Column(
-        modifier = modifier
-    ) {
-        Image(
-            painter = painterResource(id = R.drawable.img_precise_store_jyp_bubble),
-            contentDescription = null,
-            modifier = Modifier
-                .fillMaxWidth()
-                .aspectRatio(PRECISE_STORE_BANNER_IMAGE_RATIO),
-            contentScale = ContentScale.Crop
-        )
-
-        Spacer(modifier = Modifier.height(20.dp))
-
-        Text(
-            text = stringResource(id = R.string.precise_store_bubble_introduction),
-            color = White,
-            style = Body02
-        )
-
-        Text(
-            text = description,
-            color = Gray400,
-            style = Body03,
-            modifier = Modifier.padding(top = 8.dp)
-        )
-    }
-}
-
-@Composable
-private fun PreciseStoreCheckBoxes(
-    isCheckedList: List<Boolean>,
-    onClickCheckBox: (Int) -> Unit,
-    modifier: Modifier = Modifier
-) {
-    Column(
-        modifier = modifier
-    ) {
-
-        for (index in 0..<CHECK_BUTTON_NUM) {
-            with(checkBoxList[index]) {
-                PreciseStoreCheckBox(
-                    checkBoxContent = this,
-                    isChecked = isCheckedList[index],
-                    onClickCheckBox = { onClickCheckBox(index) }
-                )
-
-                if (index < CHECK_BUTTON_NUM - 1) {
-                    Spacer(modifier = Modifier.height(6.dp))
-                }
-            }
-        }
-    }
-}
 
 
 @Preview(showBackground = true)
