+++ app/src/main/java/com/sopt/bubble/feature/precise_store/component/PreciseStoreBubbleDescriptionView.kt
@@ -1,0 +1,56 @@
+package com.sopt.bubble.feature.precise_store.component
+
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.aspectRatio
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.layout.ContentScale
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.unit.dp
+import com.sopt.bubble.R
+import com.sopt.bubble.feature.precise_store.PreciseStoreViewModel.Companion.PRECISE_STORE_BANNER_IMAGE_RATIO
+import com.sopt.bubble.ui.theme.Body02
+import com.sopt.bubble.ui.theme.Body03
+import com.sopt.bubble.ui.theme.Gray400
+import com.sopt.bubble.ui.theme.White
+
+@Composable
+fun PreciseStoreBubbleDescriptionView(
+    description: String,
+    modifier: Modifier = Modifier
+) {
+    Column(
+        modifier = modifier
+    ) {
+        Image(
+            painter = painterResource(id = R.drawable.img_precise_store_jyp_bubble),
+            contentDescription = null,
+            modifier = Modifier
+                .fillMaxWidth()
+                .aspectRatio(PRECISE_STORE_BANNER_IMAGE_RATIO),
+            contentScale = ContentScale.Crop
+        )
+
+        Spacer(modifier = Modifier.height(20.dp))
+
+        Text(
+            text = stringResource(id = R.string.precise_store_bubble_introduction),
+            color = White,
+            style = Body02
+        )
+
+        Text(
+            text = description,
+            color = Gray400,
+            style = Body03,
+            modifier = Modifier.padding(top = 8.dp)
+        )
+    }
+}
